moves to correct directory
